Extract GroupedExecutionTagger to a top level class
PlanFragmenter was getting bloated and hard to read, and GroupedExectuionTagger
is complex and independent.
